pyproject.toml
[tool.poetry]
name = "conferencia de debitos erp e conta azul"
version = "0.1.0"
description = ""
authors = ["geovanecarvalho <geovaneoliveira.75@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.50.0"
unidecode = "^1.3.8"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
